found bug thanks to Firefox
