+++ nengo_gui/static/netgraph/items/item.ts
@@ -11,13 +11,10 @@
  * @param {string|null} info.parent - a NetGraphItem with .type=="net"
  */
 
-import { VNode, dom, h  } from "maquette";
-
 // import * as menu from "../../menu";
-import { Shape } from "../../utils";
 import { NetGraph } from "../netgraph";
 import { NetItem } from "./resizable";
-
+import { NetGraphItemView } from "./views/item";
 
 export interface NetGraphItemArg {
     ng: NetGraph;
@@ -28,59 +25,34 @@ export interface NetGraphItemArg {
     posX: number;
     posY: number;
 
-    dimensions: number;
     parent: string;
 
     uid: string;
 }
 
-export abstract class NetGraphItem {
+// TODO: make abstract once basics are completed
+export class NetGraphItem {
     childConnections;
     children;
     connIn;
     connOut;
     depth: number;
-    dimensions: number;
-    g: SVGElement;
-    gItems: SVGElement;
     gNetworks: SVGElement;
     menu;
-    minHeight: number;
-    minWidth: number;
     ng: NetGraph;
     parent: NetItem;
-    shape: SVGElement;
-    size: number[];
+    view: NetGraphItemView;
     uid: string;
-
-    protected _h: number;
-    protected _w: number;
-    protected _x: number;
-    protected _y: number;
-
-    protected _fuck: number;
 
     constructor(ngiArg: NetGraphItemArg) {
 
         this.ng = ngiArg.ng;
         this.uid = ngiArg.uid;
-
-        this._w = ngiArg.width;
-        this._h = ngiArg.height;
-        this._x = ngiArg.posX;
-        this._y = ngiArg.posY;
-        this.dimensions = ngiArg.dimensions;
-
         this.gNetworks = this.ng.gNetworks;
-        this.gItems = this.ng.gItems;
 
         // NetGraphConnections leading into and out of this item
         this.connOut = [];
         this.connIn = [];
-
-        // Minimum and maximum drawn size, in pixels
-        this.minWidth = 5;
-        this.minHeight = 5;
 
         // Determine the parent NetGraphItem (if any) and the nested depth
         // of this item.
@@ -88,153 +60,16 @@ export abstract class NetGraphItem {
             this.parent = null;
             this.depth = 1;
         } else {
+            // TODO: I think this is wrong
             this.parent = this.ng.svgObjects.net[ngiArg.parent];
             this.depth = this.parent.depth + 1;
         }
 
-        // Create the SVG group to hold this item's shape and it's label
-        this.g = dom.create(h("g")).domNode as SVGElement;
-        this.gItems.appendChild(this.g);
+        this.view = new NetGraphItemView(ngiArg);
 
         // TODO: find out where to render the menu
         //this.menu = new menu.Menu(this.ng.parent);
     }
-
-    get x(): number {
-        return this._x;
-    }
-
-    get y(): number {
-        return this._y;
-    }
-
-    get height(): number {
-        return this._h;
-    }
-
-    /**
-     * Return the height of the item, taking into account parent heights.
-     */
-    get nestedHeight() {
-        let h = this.height;
-        let parent = this.parent;
-        while (parent !== null) {
-            h *= parent.height * 2;
-            parent = parent.parent;
-        }
-        return h;
-    }
-
-    get fuck(): number {
-        return this._w;
-    }
-
-    set fuck(val: number) {
-        this._w = val;
-    }
-
-    get width(): number {
-        return this._w;
-    }
-
-    /**
-     * Return the width of the item, taking into account parent widths.
-     */
-    get nestedWidth() {
-        let w = this.width;
-        let parent = this.parent;
-        while (parent !== null) {
-            w *= parent.width * 2;
-            parent = parent.parent;
-        }
-        return w;
-    }
-
-    get displayedSize() {
-        return [this.width, this.height];
-    }
-
-    get displayedShape(): Shape {
-        const screenW = this.screenWidth;
-        const screenH = this.screenHeight;
-        return {height: screenH, width: screenW};
-    }
-
-    /* 
-    Different implementation for normal and mini objects 
-    */
-    abstract _getScreenW(): number;
-
-    get screenWidth() {
-        let screenW = this._getScreenW();
-
-        if (screenW < this.minWidth) {
-            screenW = this.minWidth;
-        }
-
-        return screenW * 2;
-    }
-
-    abstract _getScreenH(): number;
-
-    get screenHeight() {
-        let screenH = this._getScreenH();
-
-        if (screenH < this.minHeight) {
-            screenH = this.minHeight;
-        }
-
-        return screenH * 2;
-    }
-
-    // TODO: rename
-    abstract _getPos(): {
-        w: number, h: number, offsetX: number, offsetY: number
-    };
-
-    /**
-     * Determine the pixel location of the centre of the item.
-     */
-    get screenLocation() {
-        // FIXME: this should probably use this.ng.getScaledWidth
-        // and this.ng.getScaledHeight
-
-        const pos = this._getPos();
-
-        let dx = 0;
-        let dy = 0;
-        let parent = this.parent;
-        while (parent !== null) {
-            dx *= parent.width * 2;
-            dy *= parent.height * 2;
-
-            dx += (parent.x - parent.width);
-            dy += (parent.y - parent.height);
-            parent = parent.parent;
-        }
-        dx *= pos.w;
-        dy *= pos.h;
-
-        let ww = pos.w;
-        let hh = pos.h;
-        if (this.parent !== null) {
-            ww *= this.parent.nestedWidth * 2;
-            hh *= this.parent.nestedHeight * 2;
-        }
-
-        return [this.x * ww + dx + pos.offsetX,
-                this.y * hh + dy + pos.offsetY];
-    }
-
-    get minMaxXY() {
-        const minX = this.x - this.width;
-        const maxX = this.x + this.width;
-        const minY = this.y - this.height;
-        const maxY = this.y + this.height;
-        return [minX, maxX, minY, maxY];
-    }
-
-    abstract _renderShape();
 
     createModal() {
         this.ng.notify("createModal", {
@@ -269,19 +104,14 @@ export abstract class NetGraphItem {
         }
 
         // Remove from the SVG
-        this.gItems.removeChild(this.g);
+        this.view.remove();
         if (this.depth === 1) {
             this.ng.scaleMiniMap();
         }
     }
 
     redrawPosition() {
-        const screen = this.screenLocation;
-
-        this.g.setAttribute(
-            "transform",
-            "translate(" + screen[0] + ", " + screen[1] + ")"
-        );
+        this.view.redrawPosition();
     }
 
     redrawConnections() {
@@ -299,14 +129,5 @@ export abstract class NetGraphItem {
     redraw() {
         this.redrawPosition();
         this.redrawConnections();
-    }
-
-    /**
-     * Helper function for setting attributes.
-     */
-    setAttributes(el, attrs) {
-        Object.keys(attrs).forEach(key => {
-            el.setAttribute(key, attrs[key]);
-        });
     }
 }
